backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
chromadb>=0.5.0
httpx>=0.27.0
google-generativeai>=0.8.0
python-multipart>=0.0.9
pypdf>=4.0.0
